pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "athlete-mcp"
version = "0.1.0"
description = "MCP-powered athletic training tracker with FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "libsql-client>=0.3.1",
    "certifi>=2024.0.0",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx",
]

[project.scripts]
athlete-api = "athlete_mcp.api.main:app"
athlete-logger = "athlete_mcp.servers.logger_server:main"
athlete-analytics = "athlete_mcp.servers.analytics_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/athlete_mcp"]
